# FIX: skip empty lines in read_raw_eyelink