--- a/mne/io/eyelink/_utils.py
+++ b/mne/io/eyelink/_utils.py
@@ -122,6 +122,8 @@ def _parse_recording_blocks(fname):
                 is_recording_block = True
             if is_recording_block:
                 tokens = line.split()
+                if not tokens:
+                    continue  # skip empty lines
                 if tokens[0][0].isnumeric():  # Samples
                     data_dict["sample_lines"].append(tokens)
                 elif tokens[0] in data_dict["event_lines"].keys():